remove circular dependency on TreeNode
